+++ controllers/authController.js
@@ -2,6 +2,7 @@ const jwt = require('jsonwebtoken');
 const db = require('../config/db');
 const { verifyFirebaseToken, revokeFirebaseToken } = require('../utils/firebaseService');
 const { sendWelcomeEmail } = require('../utils/emailService');
+const { LoginDTO, LogoutDTO } = require('../dtos');
 
 const FRONT_END_URL = process.env.FRONT_END_URL || 'http://localhost:3000';
 
@@ -12,19 +13,26 @@ const FRONT_END_URL = process.env.FRONT_END_URL || 'http://localhost:3000';
 const firebaseLogin = async (req, res) => {
     const requestId = Math.random().toString(36).substring(7);
     try {
-        const { firebaseToken, rolId = 1, email, nombre, apellido, photoURL } = req.body;
+        // Usar LoginDTO para validación
+        const loginDTO = new LoginDTO(req.body);
+        const validation = loginDTO.validate();
+        
+        if (!validation.isValid) {
+            console.log(`❌ [${requestId}] Invalid login data:`, validation.errors);
+            return res.status(400).json({ 
+                error: 'Datos de login inválidos', 
+                errors: validation.errors 
+            });
+        }
+
+        const { firebaseToken, rolId, email, nombre, apellido, photoURL } = loginDTO.toFirebaseFormat();
 
         console.log(`\n📝 [${requestId}] Firebase login request:`, {
             hasToken: !!firebaseToken,
+            hasRolId: !!rolId,
             tokenLength: firebaseToken ? firebaseToken.length : 0,
             email,
-            rolId
         });
-
-        if (!firebaseToken) {
-            console.log(`❌ [${requestId}] Missing Firebase token`);
-            return res.status(400).json({ error: 'Firebase token is required' });
-        }
 
         // 1. Verificar Firebase token
         console.log(`🔐 [${requestId}] Verifying Firebase token...`);
@@ -71,6 +79,9 @@ const firebaseLogin = async (req, res) => {
 
         let userId;
         let userData;
+        let hasRol = false;
+        let requiresRoleSelection;
+        let appToken;
 
         if (users.length > 0) {
             // Usuario existente
@@ -78,32 +89,45 @@ const firebaseLogin = async (req, res) => {
             userData = users[0];
             console.log(`👤 [${requestId}] Existing user found: ID=${userId}`);
 
+            // Verificar si la cuenta está desactivada y reactivarla
+            if (users[0].is_active === false || users[0].is_active === 0) {
+                console.log(`🔄 [${requestId}] Reactivando cuenta...`);
+                await db.query('UPDATE users SET is_active = TRUE WHERE user_id = ?', [userId]);
+            }
+
             // Actualizar Firebase UID si no lo tiene
-        if (!userData.user_google_id || !userData.profile_image) {
-            console.log(`🔄 [${requestId}] Updating Firebase UID & profile...`);
-            await db.query(
-                'UPDATE users SET user_google_id = ?, profile_image = ? WHERE user_id = ?',
-                [firebaseUid, photoURL || decodedToken.picture || null, userId]
-            );
-        }
+            if (!userData.user_google_id || !userData.profile_image) {
+                console.log(`🔄 [${requestId}] Updating Firebase UID & profile...`);
+                await db.query(
+                    'UPDATE users SET user_google_id = ?, profile_image = ? WHERE user_id = ?',
+                    [firebaseUid, photoURL || decodedToken.picture || null, userId]
+                );
+            }
+            
+            // Verificar si tiene rol asignado
+            const [rolCheck] = await db.query('SELECT rol_id FROM user_rol WHERE user_id = ?', [userId]);
+            hasRol = rolCheck.length > 0;
+           // if(!hasRol) { console.log(`[${requestId}] Cuenta reactivada sin rol asignado`)}
         } else {
             // Usuario nuevo: crear
             console.log(`✨ [${requestId}] Creating new user:`, firebaseEmail);
-        const [result] = await db.query(
-                `INSERT INTO users (user_name, user_lastname, user_email, user_google_id, profile_image)
-                 VALUES (?, ?, ?, ?, ?)`,
-                [nombre || firebaseEmail.split('@')[0], apellido || '', firebaseEmail, firebaseUid, photoURL || decodedToken.picture || null]
-            );
-            
+            const [result] = await db.query(
+                `INSERT INTO users (user_name, user_lastname, user_email, user_google_id, profile_image, is_active)
+                 VALUES (?, ?, ?, ?, ?, TRUE)`,
+                 [nombre || firebaseEmail.split('@')[0], apellido || '', firebaseEmail, firebaseUid, photoURL || decodedToken.picture || null]
+             );
+
             userId = result.insertId;
             console.log(`✅ [${requestId}] New user created: ID=${userId}`);
-            
-            // Asignar rol
-            await db.query(
-                `INSERT INTO user_rol (user_id, rol_id, start_date) VALUES (?, ?, NOW())`,
-                [userId, rolId]
-            );
-            
+
+            if (rolId) {
+                await db.query(
+                    `INSERT INTO user_rol (user_id, rol_id, start_date) VALUES (?, ?, NOW())`,
+                    [userId, rolId]
+                );
+                hasRol = true;
+            }
+
             // Enviar correo de bienvenida (no bloquea el registro si falla)
             sendWelcomeEmail(
                 firebaseEmail,
@@ -112,30 +136,36 @@ const firebaseLogin = async (req, res) => {
             ).catch(err =>
                 console.error(`❌ [${requestId}] Error enviando correo de bienvenida:`, err.message || err)
             );
-            
+
             // Obtener datos del usuario creado
             [users] = await db.query(
                 `SELECT U.user_id, U.user_name, U.user_lastname, U.user_email,
-                        U.user_phonenumber, U.whatsapp, UR.rol_id
+                        U.user_phonenumber, U.whatsapp, U.profile_image, UR.rol_id
                  FROM users AS U
                  LEFT JOIN user_rol AS UR ON U.user_id = UR.user_id
-                 WHERE U.user_id = ?
-                 LIMIT 1`,
+                 WHERE U.user_id = ?`,
                 [userId]
             );
             userData = users[0];
         }
 
         // 3. Generar JWT de la app (no el de Firebase)
-        console.log(`🔑 [${requestId}] Generating app JWT...`);
-        const appToken = jwt.sign(
-            { id: userId, rol: userData.rol_id || rolId },
-            process.env.JWT_SECRET,
-            { expiresIn: process.env.JWT_EXPIRES || '24h' }
-        );
+        if(hasRol) {
+            console.log(`🔑 [${requestId}] Generating app JWT...`);
+            appToken = jwt.sign(
+                { id: userId, rol: userData.rol_id || rolId },
+                process.env.JWT_SECRET,
+                { expiresIn: process.env.JWT_EXPIRES || '24h' }
+            );
+            requiresRoleSelection = false;
+        } else {
+            console.log(`[${requestId}] First Login - require role selection`);
+            appToken = null;
+            requiresRoleSelection = true;
+        }
 
         // 4. Devolver datos del usuario y token
-    let userPayload = {
+        let userPayload = {
             id: userId,
             nombre: userData.user_name,
             apellido: userData.user_lastname,
@@ -143,16 +173,17 @@ const firebaseLogin = async (req, res) => {
             profile_image: userData.profile_image,
             telefono: userData.user_phonenumber || null,
             whatsapp: userData.whatsapp || null,
-            rol: userData.rol_id || rolId,
-            token: appToken,
+            photoURL: userData.profile_image,
+            rol: userData.rol_id || null,
         };
 
-        console.log(`✅ [${requestId}] Firebase login successful:`, { userId, email: userData.user_email });
+        console.log(`✅ [${requestId}] Firebase login successful:`, { userId, requiresRoleSelection, hasRol });
 
         res.json({
             success: true,
             user: userPayload,
             token: appToken,
+            requiresRoleSelection
         });
 
     } catch (error) {
@@ -198,12 +229,19 @@ const githubCallback = (req, res) => {
  */
 const logout = async (req, res) => {
     try {
-        const { firebaseToken } = req.body;
+        // Usar LogoutDTO para validación
+        const logoutDTO = new LogoutDTO(req.body);
+        const validation = logoutDTO.validate();
         
-        if (!firebaseToken) {
-            return res.status(400).json({ error: 'Firebase token is required' });
+        if (!validation.isValid) {
+            return res.status(400).json({ 
+                error: 'Datos de logout inválidos', 
+                errors: validation.errors 
+            });
         }
 
+        const { firebaseToken } = logoutDTO;
+        
         const decodedToken = await verifyFirebaseToken(firebaseToken);
         await revokeFirebaseToken(decodedToken.uid);
