+++ dtos/index.js
@@ -1,18 +1,30 @@
 /**
  * Exportación centralizada de todos los DTOs
  */
+const { UserDTO, CreateUserDTO, UpdateUserDTO } = require('./UserDTO');
+const { ApartmentDTO, CreateApartmentDTO, UpdateApartmentDTO } = require('./ApartmentDTO');
+const { ReviewDTO, CreateReviewDTO, UpdateReviewDTO } = require('./ReviewDTO');
+const { ContractDTO, CreateContractDTO, UpdateContractDTO } = require('./ContractDTO');
+const { NotificationDTO, CreateNotificationDTO, UpdateNotificationDTO } = require('./NotificationDTO');
+const { LoginDTO, RefreshTokenDTO, LogoutDTO } = require('./AuthDTO');
+
 module.exports = {
-    UserDTO: require('./UserDTO'),
-    CreateUserDTO: require('./CreateUserDTO'),
-    UpdateUserDTO: require('./UpdateUserDTO'),
-    ApartmentDTO: require('./ApartmentDTO'),
-    CreateApartmentDTO: require('./CreateApartmentDTO'),
-    UpdateApartmentDTO: require('./UpdateApartmentDTO'),
-    ReviewDTO: require('./ReviewDTO'),
-    CreateReviewDTO: require('./CreateReviewDTO'),
-    UpdateReviewDTO: require('./UpdateReviewDTO'),
-    CreateContractDTO: require('./CreateContractDTO'),
-    UpdateContractDTO: require('./UpdateContractDTO'),
-    NotificationDTO: require('./NotificationDTO'),
-    CreateNotificationDTO: require('./CreateNotificationDTO')
+    UserDTO,
+    CreateUserDTO,
+    UpdateUserDTO,
+    ApartmentDTO,
+    CreateApartmentDTO,
+    UpdateApartmentDTO,
+    ReviewDTO,
+    CreateReviewDTO,
+    UpdateReviewDTO,
+    ContractDTO,
+    CreateContractDTO,
+    UpdateContractDTO,
+    NotificationDTO,
+    CreateNotificationDTO,
+    UpdateNotificationDTO,
+    LoginDTO,
+    RefreshTokenDTO,
+    LogoutDTO
 };
