+++ routes/auth.js
@@ -1,7 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const { verifyToken, generateAccessToken } = require('../utils/auth');
-const { firebaseLogin, githubRedirect, githubCallback } = require('../controllers/authController');
+const { firebaseLogin, githubRedirect, githubCallback, logout } = require('../controllers/authController');
+const { RefreshTokenDTO } = require('../dtos');
 
 // Firebase Google Sign-In: recibe Firebase token y devuelve JWT de la app
 router.post('/firebase-login', firebaseLogin);
@@ -19,10 +20,19 @@ router.get('/github/callback', githubCallback);
 
 // Renovar access token con refresh token
 router.post('/refresh-token', (req, res) => {
-    const { refreshToken } = req.body;
-    if (!refreshToken) {
-        return res.status(401).json({ error: 'Token de refresco no proporcionado' });
+    // Usar RefreshTokenDTO para validación
+    const refreshDTO = new RefreshTokenDTO(req.body);
+    const validation = refreshDTO.validate();
+    
+    if (!validation.isValid) {
+        return res.status(401).json({ 
+            error: 'Token de refresco inválido', 
+            errors: validation.errors 
+        });
     }
+    
+    const { refreshToken } = refreshDTO;
+    
     try {
         const decoded = verifyToken(refreshToken, process.env.JWT_REFRESH_SECRET);
         const newAccesToken = generateAccessToken({ id: decoded.id, role: decoded.role });
@@ -35,10 +45,19 @@ router.post('/refresh-token', (req, res) => {
 
 // Alias para /refresh
 router.post('/refresh', (req, res) => {
-    const { refreshToken } = req.body;
-    if (!refreshToken) {
-        return res.status(401).json({ error: 'Token de refresco no proporcionado' });
+    // Usar RefreshTokenDTO para validación
+    const refreshDTO = new RefreshTokenDTO(req.body);
+    const validation = refreshDTO.validate();
+    
+    if (!validation.isValid) {
+        return res.status(401).json({ 
+            error: 'Token de refresco inválido', 
+            errors: validation.errors 
+        });
     }
+    
+    const { refreshToken } = refreshDTO;
+    
     try {
         const decoded = verifyToken(refreshToken, process.env.JWT_REFRESH_SECRET);
         const newAccesToken = generateAccessToken({ id: decoded.id, role: decoded.role });
@@ -49,16 +68,6 @@ router.post('/refresh', (req, res) => {
     }
 });
 
-router.get('/logout', (req, res) => {
-    req.logout(() => {
-        req.session.destroy((err) => {
-            if (err) {
-                console.error('Error cerrando sesión:', err);
-                return res.status(500).json({ error: 'No se pudo cerrar sesión' });
-            }
-            res.json({ message: 'Sesión cerrada correctamente' });
-        });
-    });
-});
+router.post('/logout', logout);
 
 module.exports = router;
